netdog: Rename device-specific "bonding" module to "bond"
`net_config` has a "devices" module that contains the config structures
for interfaces, vlans, and bonds.  The module for bond devices was
previously named "bonding" which didn't conform to the other modules
which had names that were nouns (interface, vlan).  This commit renames
the `devices/bonding` module to `devices/bond`.
