+++ sources/api/netdog/src/net_config/devices/mod.rs
@@ -2,13 +2,13 @@
 //! intended to be the structs used for net.toml deserialization including the validation logic for
 //! each device.
 
-pub(crate) mod bonding;
+pub(crate) mod bond;
 pub(crate) mod interface;
 pub(crate) mod vlan;
 
 use super::{error, Result, Validate};
 use crate::addressing::{Dhcp4ConfigV1, Dhcp6ConfigV1, RouteV1, StaticConfigV1};
-use bonding::NetBondV1;
+use bond::NetBondV1;
 use interface::NetInterfaceV2;
 use serde::Deserialize;
 use vlan::NetVlanV1;
